refactor: MainWindowUI.initMainMenu
Move initMainMenu static method from Core to MainWindowUI utility class.
